requirements.txt
fastapi==0.42.0
ujson==1.35
uvicorn==0.10.3
hydra-core==0.10.0
httpx==0.7.6
